+++ finagle-zipkin-core/src/main/scala/com/twitter/finagle/zipkin/core/DeadlineSpanMap.scala
@@ -63,8 +63,6 @@ private class DeadlineSpanMap(
         // it's already been flushed. copy most of it and immediately log it.
         val copy = ms.copyForImmediateLogging()
         f(copy)
-        copy.addAnnotation(
-          ZipkinAnnotation(Time.nowNanoPrecision, "finagle.zipkin.late_arrival", ms.endpoint))
         Some(copy)
       } else {
         f(ms)
@@ -96,7 +94,6 @@ private class DeadlineSpanMap(
       val ms = iter.next().getValue
       if (ms.started <= deadline) {
         ms.synchronized {
-          ms.addAnnotation(ZipkinAnnotation(deadline, "finagle.flush", ms.endpoint))
           beforeDeadline.append(ms.toSpan)
           preventMoreUpdates(ms)
           iter.remove()